requirements.txt
fastapi
uvicorn
motor
pydantic
google-generativeai
python-dotenv
python-jose
passlib[bcrypt]
streamlit
requests 